Adds WAF rules to terraform provider with documentation
Note that since WAF rules already exist for an account we're not
really "creating" them, only pointing to them and storing their
state.

Since the ID is known at creation time and we need to "point"
to the actual WAF rule, I've added "rule_id" as an attribute that
the user can set based on the ID they can see in the UI.

An example configuration looks like this:

```
resource "cloudflare_waf_rule" "100000" {
  rule_id = "100000"
  zone = "domain.com"
  mode = "simulate"
}
```

The create function will look up the WAF rule 100000, update
it's state depending on the mode passed in by the user, then
store the result.

The read function checks for any updates to the mode of the WAF rule.

The update function attempts to modify the mode of a WAF rule.

The delete function attempts to set the WAF rule to its default
mode.

Import is also implemented and tested in the convention of other
import test files in the repo.