Remove unnessecary `ForceNew` attributes
`ForceNew` should be used where a resource cannot be updated in place
and requires a fresh start. This is currently the case for
`cloudflare_load_balancer_pool` as it doesn't implement an update
function however that will be addressed in a future change but first, we
need to remove this hurdle.