Support load balancer origin weights and session affinity (#153)
* Support weight parameter for load balancer origins

Updating the cloudflare-go library brings support for the weight
parameter of load balancer origins. This allows configuration of traffic
priority to each of the backend origins, which is critical in more
complex load balancing setups.

* Support session_affinity for load balancers

The newer cloudflare-go library supports configuring session affinity
values for Cloudflare load balancers. This commit updates the terraform
resource to support this field.