Ensure `zone` and `zone_id` are always set (#175)
* Ensure `zone` and `zone_id` are always set

During an import of the `cloudflare_access_rule` resource, the `zone_id`
is defined within the composite ID. This value is then populated and
synced within `resourceCloudflareAccessRuleRead` to match the expected
schema resource. However, we don't currently set zone (the zone name)
anywhere. This becomes problematic on subsequent terraform operations as
the Read attempts to sync the schema with the access rule and ends up
attempting to recreate the resource due to detecting a change.

This is a similar issue to #161 and is addressed in a similar manner.

Fixes #166