custom_page: Define `SetID` to save state
Prior to this change we weren't setting the `Id` value for these
resources when creating them from new. If you imported, it worked as the
composite ID (`customPageLevel`/`identifier`/`pageType`) was defined.

To fix the issue, we're calling `d.SetId()` and setting the ID
explicitly as a md5 checksum of the `identifier` and `pageType`. This is
in order to maintain consistency with the other resources which use a 32
character ID.

Fixes #260