data_source_zone: Return ID and Name
Updates the `cloudflare_zones` data source to return a map of the ID and
name for the filtered results. This is to better help facilitate dynamic
configurations where end users would like to pass in a zone filter and
automate creation of resources.

There are a couple of potential downsides though:

- The first is that this introduces a somewhat breaking change as it
  doesn't perform as it once did. An entry will be added to the CHANGELOG
  but it could be easy to miss.
- The second is that the syntax to access a single result is pretty
  ugly. This will be nicer in 0.12 but for now it's 🤮.

Closes #211