Bump Terraform to 0.12.0-beta1
This commit bumps Terraform to work with the new version of Terraform.

It also cleans up the `go.mod` file. Now that our dependencies are also
using `go mod`, Go won't populate the `// indirect` on the file anymore.

There is a small caveat to compile the transitive dependencies without
needing to install `bazaar`, or having it failing to work on Windows. We
need to replace them with alternative sources that are not on Bazaar.