cloudflare_page_rule: Allow `browser_cache_ttl` of 0
In c1799f4, @patryk made some improvements on how the `Update` handles
working out which rules need to be changed. With the introduction of
that change, the `transformToCloudflarePageRuleAction` now accepts a
third argument of whether the action has changed. For the majority of
rules, this isn't problematic however it has caused issues with
`browser_cache_ttl` as a value of 0 tells Cloudflare to respect the
origin cache control headers however, here it would end up setting the
page rule action value to `nil` (as it hadn't changed and was equal to
0).

To start with, I was going to remove the `intValue == 0` portion of the
check and just rely on whether the rule had been changed but this didn't
work due to the way disabling `edge_cache_ttl` works (it is set to 0
for removal). Instead, I moved the check into the conditional for the
`pageRuleAPIFloatFields` `elseif` which will allow the functionality for
both to work actions to work as desired.