Remove tagged release from module
We've not seen an update from RenovateBot on the `cloudflare-go`
dependency and I suspect it's due to having a version in the module file
that is older than the newer releases. This requires that we manually
run `go get ...` to update the dependency instead of it being
automatically done.

To work around this, I've manually updated the release to be `0.0.0`
(but the same commit) which should trigger the update for RenovateBot.

This isn't the ideal as we _should_ be using tagged releases however
that is something that still needs to be solved upstream in
cloudflare/cloudflare-go#271.