Define `zoneID` for spectrum application import
The `spectrum_application` resource uses the composite ID of `zoneID`
and `spectrumApplicationID` to import the resource. Prior to this
change it was actually passing in the zone name which wasn't getting
routed correctly in the API, resulting in the following failure.

```
import cloudflare_spectrum_application.ydysyld286 cloudflare_spectrum_application.ydysyld286: Error reading spectrum application resource from API for resource hashicorptest.com in zone d984fd1105204f75a774901393dc245a: error from makeRequest: HTTP status 400: content "{\"success\":false,\"errors\":[{\"code\":7003,\"message\":\"Could not route to \\/zones\\/hashicorptest.com\\/spectrum\\/apps\\/d984fd1105204f75a774901393dc245a, perhaps your object identifier is invalid?\"},{\"code\":7000,\"message\":\"No route for that URI\"}],\"messages\":[],\"result\":null}"
```

To remedy this, we need to pass in the correct zone ID for the zone we
are attempting to operate on. In the tests, this isn't easy and instead
of trying to expose the Cloudflare library to the test suite, I've added
an environment lookup with a default value of the primary test zone.

As mentioned in the code comment, this is only temporary. Once #288 is
complete and `CLOUDFLARE_ZONE_ID` is defined in CI, we can just have it
read from that and remove the default. We will want to add in a precheck
for the new variable but that can be added in another PR.