Add sweeper for `cloudflare_zones`
Creates a Terraform sweeper[1] to clean up any dangling resources. Here,
for Cloudflare zones that are being used a test for management via the
API. I've restricted the lookup to the ephemeral zones we want to
remove instead of all of them as I suspect removing all would make the
other tests fail that depend on it always being there.

Closes #296

[1]: https://www.terraform.io/docs/extend/testing/acceptance-tests/sweepers.html