filters: ignore \n, \t and whitespaces at the beginning and end of expressions
This commit fixes an issue where filters using the heredoc syntax are always
seen as "to change" by Terraform. This is due to the fact that the heredoc
syntax includes a trailing character (\n) before EOF and because Cloudflare
always strips these trailing characters before storing expressions, leading
to Terraform always thinking that something must be changed. This change
also mimic the behavior of Cloudflare that always strips carriage-return,
whitespaces and tabs at the beginning and the end of expressions.