test/spectrum_application: Add test sweeper
Recently we have been seeing some test flakiness for spectrum
applications due to left over resources or partially deleted resources.
To combat this, we can use a test sweeper (similar to what we've done
with `zones`) which will ensure a clean test bed for all runs.

This fixes the following CI failure.

```
------- Stdout: -------
=== RUN   TestAccCloudflareSpectrumApplication_Basic
--- FAIL: TestAccCloudflareSpectrumApplication_Basic (1.46s)
    testing.go:568: Step 0 error: errors during apply:

        Error: error creating spectrum application for zone: error from makeRequest: HTTP status 400: content "{\n  \"result\": null,\n  \"success\": false,\n  \"errors\": [\n    {\n      \"code\": 12004,\n      \"message\": \"Failed to associate application: Port(s) conflict with existing application.\"\n    }\n  ],\n  \"messages\": []\n}\n"
```