resource/cloudflare_zone: Update test domain
CI has recently started failing the `TestAccCloudflareZone` test as the
`example.org` domain is using old Cloudflare nameservers and the work to
address this is on a backlog to be fixed. Instead of using an example
domain, we can swap to using an ephemeral subdomain and perform the
assertions against that.