provider: Rename "token" => "key"
API "token" is currently used for HTTP authentication requests where the
`X-Auth-Key` needs to be set. Until recently, it was fine to use the
wrong term for this authentication method (as there was only the one)
however there is a new feature upcoming from Cloudflare which is
actually called API tokens and it requires a different authentication
procedure.

This means we need to rename the existing variables and references to
allow us to build in the new functionality when it is announced.

I've updated all the references to use the term "key" as that is what
the documentation and values are and I think it's a better reflection of
what method is in use than what we have in place at the moment.