Deprecation for `zone` in favour of explicit `zone_id`
Zone names within Cloudflare are not universally unique which makes
performing a lookup based on just the name a little troublesome. Two
accounts can have the same zone name and both of these can be returned
if the lookups are not restricted to the zone ID or account ID.

To rememdy this, `zone` is being deprecated in favour of explicit
`zone_id` values. See the discussion[1] on this for full details.

Following Terraform's guidelines for deprecations[2], we are putting in
a deprecation notice as part of a minor release and then in our next
major release, we will actually perform the schema and functionality
removal. This update will not prevent people from using `zone` however
use will trigger a warning that it's going away and they should migrate
themselves.

[1]: https://github.com/terraform-providers/terraform-provider-cloudflare/issues/161#issuecomment-441128985
[2]: https://www.terraform.io/docs/extend/best-practices/deprecations.html#provider-attribute-removal