page_rule: Remove `HasChange` for `edge_cache_ttl`
Prior to swapping to PUT requests for updating page rules, we used to
perform manual diff checks that would determine which actions we would
add to the page rule before sending it. This is no longer now as we
swapped to using a PUT request which replaces the whole rule and we are
able to send the whole page rule as-is without doing the work ourselves.

This popped up as it's caused a bug whereby updating a page rule that
includes a `edge_cache_ttl` (but not change it) results in the
`edge_cache_ttl` being dropped completely.

`TestAccCloudflarePageRuleEdgeCacheTTLNotClobbered` is the regression
test case that demonstrates the behaviour.