SDK-80: Zone to zone_id refactoring (#472)
* Refactor zone_id for resource_cloudflare_access_rule

* Refactor zone_id for cloudflare_filter

* Refactor zone_id for cloudflare_firewall_rule

* Refactor zone_id for cloudflare_load_balancer

* Refactor zone_id for cloudflare_page_rule

* Refactor zone_id for cloudflare_rate_limit

* Refactor zone_id for cloudflare_record

* Refactor zone_id for cloudflare_waf_rule

* Refactor zone_id for cloudflare_worker_route

* Refactor zone_id for cloudflare_worker_script

* Refactor zone_id for cloudflare_zone_lockdown

* Refactor zone_id for cloudflare_zone_settings_override

* Make tests `zone_id` compatible

Gets the build green by updating all tests to match the expected
behaviour of using zone ID instead of relying on zone from the
schema.

* Get rid of additional test output

* Add CLOUDFLARE_ALT_ZONE_ID to precheck

* Don't use API token authentication for Workers

The Workers service does not currently support the use of API tokens
(which is used by default when the `CLOUDFLARE_API_TOKEN` environment
variable is set[1]) and throws a very misleading exception.

```
=== RUN   TestAccCloudflareWorkerScript_SingleScriptNonEnt
--- FAIL: TestAccCloudflareWorkerScript_SingleScriptNonEnt (0.59s)
    testing.go:569: Step 0 error: errors during apply:

        Error: script already exists.

          on /var/folders/d4/5sgps61s2jg8f0_71663vw800000gn/T/tf-test617106443/main.tf line 2:
          (source code not available)
```

The reason we hit this exception is due to the `WorkerScript.Script`
being populated with the API error instead of being empty which falls
into this `if` block[2].

```
(cloudflare.WorkerScript) {
 WorkerMetaData: (cloudflare.WorkerMetaData) {
  ID: (string) "",
  ETAG: (string) "",
  Size: (int) 0,
  CreatedOn: (time.Time) 0001-01-01 00:00:00 +0000 UTC,
  ModifiedOn: (time.Time) 0001-01-01 00:00:00 +0000 UTC
 },
 Script: (string) (len=105) "{\"success\":false,\"errors\":[{\"code\":10000,\"message\":\"API Tokens are not supported by this API for now\"}]}\n"
}
```

This was initially puzzling as `err` is handled in `cloudflare-go` where
this field is defined[3] but looking deeper, the HTTP response is coming
back as a 200 OK response, hence the `Script` value being populated
and not falling into the `err` handling.

To fix this, it is two fold. The first is unsetting the
`CLOUDFLARE_API_TOKEN` environment variable in CI. This gets the tests
passing again however to address the underlying issue, we've contacted
Cloudflare to fix the HTTP response code which will then made everything
in the library work as intended. Depending on the length of time before
the fix is deployed, we might also hotfix the library to instead check
`r.Success = true` as well to fail in the correct spot instead of
allowing this to bubble up in unexpected ways.

[1]: https://git.io/Jes9l
[2]: https://git.io/Jes9A
[3]: https://git.io/Jes9h