resource/cloudflare_zone: account for pending zone setups
During the creation of a zone, the plan name is always "free" despite
the desired end result of a paid plan while it is in a "pending" state.
In most API scenarios this doesn't really matter as the plan value
shifts on it's own behind the scenes. However as Terraform keeps track
of the input vs the remote state, this becomes problematic as it sees
the difference in the desired state.

To account for this, we need to check if the state is pending and then
dig into the `plan_pending` object for the real plan in order to do the
comparison.

Fixes #488