Refactor `setRatePlan` for new usage
In cloudflare/cloudflare-go#362 we are addressing an underlying issue
with an existing method used to set the zone rate plan. To move to this,
we update our usage of the library which as a side benefit, gets rid of
some of the extra work we were doing in this method.