Account for pending zone setups (#511)
* resource/cloudflare_zone: account for pending zone setups

During the creation of a zone, the plan name is always "free" despite
the desired end result of a paid plan while it is in a "pending" state.
In most API scenarios this doesn't really matter as the plan value
shifts on it's own behind the scenes. However as Terraform keeps track
of the input vs the remote state, this becomes problematic as it sees
the difference in the desired state.

To account for this, we need to check if the state is pending and then
dig into the `plan_pending` object for the real plan in order to do the
comparison.

Fixes #488

* No need to use `d.Get`, we have an object already

* Defer setting the plan type until `Read`

At the point of creation, we set the rate plan however we don't need to
worry about the status check until _after_ we've set it. This gets
re-performed in the `Read` method after the resource is created.

* Refactor `setRatePlan` for new usage

In cloudflare/cloudflare-go#362 we are addressing an underlying issue
with an existing method used to set the zone rate plan. To move to this,
we update our usage of the library which as a side benefit, gets rid of
some of the extra work we were doing in this method.

* bump cloudflare-go to v0.10.6

* Swap to backoff and retry for rate plan propagation

* Handle creates and updates

For the subscription API, creates use a POST and updates use PUT. We
also can't just POST everything so we need to determine when we are
updating and use a similar (yet slightly different) HTTP verb.