sweeper/zones: Sweep all zones (#522)
When we initially put the zone sweeper into place, it was in a shared
account that we couldn't just remove everything. This meant the
`ListZones` needed to only look up the zones we were expecting to
remove and clear those out. This is no longer the case as the tests run
against a dedicated account so we can blow everything away that isn't
the static `cfapi.net` domains (`terraform` and `terraform2`).

Fixes the following integration test failure.

```
=== RUN   TestAccCloudflareZoneBasic
--- FAIL: TestAccCloudflareZoneBasic (0.79s)
    testing.go:569: Step 0 error: errors during apply:

        Error: Error creating zone "example.cfapi.net": error from makeRequest: HTTP status 400: content "{\"success\":false,\"errors\":[{\"code\":1061,\"message\":\"example.cfapi.net already exists\"}],\"messages\":[],\"result\":null}"

          on /opt/teamcity-agent/temp/buildTmp/tf-test007851057/main.tf line 2:
          (source code not available)

FAIL
```