resource/cloudflare_zone: No need to translate the ID to Name back to ID
Within the `Update` we set the rate plan which requires the
subscription ID mapped value[1]. We already have the legacy ID[2] so we
don't need to pass it through the `planIDForName` method which we
previously.

I suspect this was a copy-paste issue from the `Read` method as we
were also setting the `plan` value to the plan name which would
haven't have been accepted in the schema anyway. This commit trims
all of that out and just passes the ID as it should have always done.

[1]: https://github.com/terraform-providers/terraform-provider-cloudflare/blob/master/cloudflare/resource_cloudflare_zone.go#L35-L40
[2]: https://github.com/terraform-providers/terraform-provider-cloudflare/blob/master/cloudflare/resource_cloudflare_zone.go#L18-L23
[3]: https://github.com/terraform-providers/terraform-provider-cloudflare/blob/master/cloudflare/resource_cloudflare_zone.go#L81