datasource/zones: Perform filtering on server side
The initial implementation of the datasource had the provider request all the
available zones and then we would perform some logic within the provider to
filter down to the match the attributes. This is slightly inefficient for a
couple of reasons:

- We are requesting *all* zones (even if we only want a small subset of them); and
- We are duplicating the zone filtering logic in the provider that is
  already available in the API

This change replaces that functionality by pushing the responsibility of
filtering back to the API. It maintains a reasonable amount of backwards
compatibility outlined in the initial PR (#168) where the "name" had to accept
wildcards but under the hood accepted a full regex. Instead, we use the
`contains:` modifier on the name search parameter which will instead fuzzy
match. The rest remains the same but is performed via the API.

Fixes #707