test/spectrum_application: Remove zone listing calls
Since we updated the `cloudflare_zone` datasource #708 to perform
different types of matching, this test has been failing due to not
having the ability to list all zones. This isn't really a problem but it
did identify we could be lazier with the resource definition and pass in
a static zone ID instead of relying on an extra zone listing call to
fetch all the information.