resource/waf_package: Only modify anomaly resources
Within WAF packages, there are a few types of resources based on their
detection mode; user, anomaly and traditional. However, only packages
with the detection mode set to "anomaly" are able to be modified by the
end user. Attempting to update the others results in the following API
exception.

```
{
  "success": false,
  "errors": [
    {
      "code": 1027,
      "message": "Only anomaly WAF packages can have their package attributes updated"
    }
  ],
  "messages": [],
  "result": null
}
```

This issue surfaces when you attempt to manage non-anomaly rule types in
Terraform as the existing `Default` values set on the schema would
differ from the assumption that all packages had these properties and it
would attempt to get them into sync.

Instead, we need to conditionally apply updates to anomaly packages and
leave the other types to be managed as a read only source. To do this,
we can look at the `detection_type` and restrict the `Update` calls to
the anomaly based WAF packages.

Fixes #723