Compare firewall descriptions after converting unicode + HTML entities
For a while now, I've noticed that we have an occasional flapping test
in our internal test suite where "example 'string'" is attempting to
trigger an update despite not actually doing anything. Diving further
into this, I found that the Cloudflare API returns unicode + HTML
entities for the descriptions.

```
{
  "result": {
    "id": "e97a2db142ae4ff09499bb6a458981fd",
    "paused": true,
    "description": "this is a \u0026#39;test\u0026#39;",
    "expression": "(http.host eq \"example.com\")"
  },
  "success": true,
  "errors": [],
  "messages": []
}
```

However most people (ok, no one) is actually wanting to write the
description that way. Instead, they are using the single characters such
as an apostrophe (`'`).

```
resource "cloudflare_filter" "%[1]s" {
  description = "this is a 'test'"
  expression = "(http.host eq \"example.com\")"
}
```

This means that Terraform constantly thinks that it needs to update the
value when it doesn't really.

To combat this, I've added a `DiffSuppressFunc` to the descriptions
which ensure the string values are compared after being processed by
`html.UnescapeString`. This accounts for the differences in the strings
and will still ensure we aren't missing updates -- even if someone
decides to actually use entities in their descriptions.