Allow retries on DNS records if the record already exists (#773)
Despite best efforts, occasionally Terraform's API calls will be faster
than the remote API's consistency. The scenario is easiest to find when
you are constantly updating similar records back and forth or changing
between A/CNAME and there is a `Delete` operation involved.

This results in a scenario where an end user will get failures from
`terraform apply` runs that indicate the record alredy exists but
moments later, is available for use.

To handle this scenario a bit nicer for the user, we can attempt to
automatically retry the operation for a short period of time. I've opted
for 30s as that is high enough that the remote API state will be in sync
but also low enough that you won't be waiting too long to find out there
is an issue there if the state is as intended and you're introducing an
unintended duplicate.

Fixes #750