resource/healthcheck: Handle deletion outside of Terraform
When a resource is managed by Terraform, any changes outside of the
configuration, should be synced to keep the state in sync. This
previously wasn't the case with healthchecks as the 404 error (here,
resource has been deleted outside of Terraform) was returning a generic
error for any non-success statuses. To fix this, we inspect the returned
error and check for a string which denotes the resource not longer
exists and then sets `d.Id()` to an empty string and Terraform handles
the lifecycle correctly.

Fixes #785