resouce/ip_list: Remove `item.id` from schema
Despite the `item` ID being marked as Computed, it is showing
unnecessary changes when running `apply` without any changes. For
whatever reason, the ID being returned is changing and Terraform is
trying to keep it in sync.

As we aren't using it, let's get rid of it and avoid the hassle of
keeping it in sync.