zone: Remove `DiffSuppressFunc` causing `jump_start` issues
Updates the `cloudflare_zone` schema to remove `DiffSuppressFunc` for
`jump_start`. From the test case in #826, this attribute is causing the
value to always be set to "false" which is preventing the end users
value from being passed through to the API call.

The history suggests this has always been the case since the resource
was introduced so I'm not sure if it was intentional, however this
addresses the bug in question.

Fixes #826