filter: Improve validation of expression
While making changes to filters you need to run `apply` to actually find
out whether the expression is valid. This is a slow feedback cycle for
developers and there is an API endpoint that allows you to validate
expressions before using them so it's a no brainer to improve the
validation in the schema.

This updates the `ValidateFunc` for `expression` to call out to the
validation API endpoint and raise those exceptions earlier in the
development cycle.

Fixes #846