filter: Propagate schema validation error to end user
#848 introduced better validation for filters using the Terraform
schema. Despite handling the error, the error wasn't returned to the end
user and instead the RPC exception is shown the end user.

```
Error: rpc error: code = Unavailable desc = transport is closing
```

This fixes the user experience by returning the errors which now passes
the exception onto the end user instead of swallowing it or causing a
panic.