Revert schema based filter validation
Due to the endpoint not supporting API tokens, we needed to build a
standalone client which hasn't worked out well and the workarounds to
fix it are against the grain of how the Terraform schema is intended to
be used leaving us in situation where change in upstream functionality
may force us to be on an older version of the SDK if we were to
continue. Instead, this reverts #848 and #860 and once API token support
is added to the endpoint, we can reintroduce this without a standalone
client.

Closes #861