resource/account_member: Use `TypeSet`, not `TypeList`
When I created this resource, `TypeList` seemed to be the correct choice
for the data type we store. However, after a while we discovered an
issue whereby a change in the ordering of the `role_ids`, resulted in the
account member wanting to be recreated. This was addressed in #128
however it meant that new roles had to be added to the end otherwise it
would flap. It also had the extra legwork of a custom `DiffSuppressFunc`
to ensure that we compared the values using a consistent ordering to
avoid the trap where the value wasn't added to the end.

Now that we understand the internal types a little better, `TypeSet` is
a more appropriate choice as it is an unordered list which fits the
`role_ids` perfectly and solves the initial issues we encountered.